Move ImageFormat to api package
