+++ jib-core/src/main/java/com/google/cloud/tools/jib/api/ImageFormat.java
@@ -14,28 +14,14 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.image;
+package com.google.cloud.tools.jib.api;
 
-import com.google.cloud.tools.jib.image.json.BuildableManifestTemplate;
-import com.google.cloud.tools.jib.image.json.OCIManifestTemplate;
-import com.google.cloud.tools.jib.image.json.V22ManifestTemplate;
-
-/** Enumeration of {@link BuildableManifestTemplate}s that indicates the format of the image. */
+/** Indicates the format of the image. */
 public enum ImageFormat {
 
   /** @see <a href="https://docs.docker.com/registry/spec/manifest-v2-2/">Docker V2.2</a> */
-  Docker(V22ManifestTemplate.class),
+  Docker,
 
   /** @see <a href="https://github.com/opencontainers/image-spec/blob/master/manifest.md">OCI</a> */
-  OCI(OCIManifestTemplate.class);
-
-  private final Class<? extends BuildableManifestTemplate> manifestTemplateClass;
-
-  ImageFormat(Class<? extends BuildableManifestTemplate> manifestTemplateClass) {
-    this.manifestTemplateClass = manifestTemplateClass;
-  }
-
-  public Class<? extends BuildableManifestTemplate> getManifestTemplateClass() {
-    return manifestTemplateClass;
-  }
+  OCI
 }
