+++ jib-core/src/test/java/com/google/cloud/tools/jib/api/JibContainerBuilderTest.java
@@ -26,7 +26,8 @@ import com.google.cloud.tools.jib.configuration.credentials.Credential;
 import com.google.cloud.tools.jib.configuration.credentials.CredentialRetriever;
 import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.event.JibEvent;
-import com.google.cloud.tools.jib.image.ImageFormat;
+import com.google.cloud.tools.jib.image.json.OCIManifestTemplate;
+import com.google.cloud.tools.jib.image.json.V22ManifestTemplate;
 import com.google.cloud.tools.jib.registry.credentials.CredentialRetrievalException;
 import com.google.common.base.Suppliers;
 import com.google.common.collect.ImmutableMap;
@@ -193,8 +194,7 @@ public class JibContainerBuilderTest {
 
     Assert.assertEquals("jib-core", buildConfiguration.getToolName());
 
-    Assert.assertSame(
-        ImageFormat.Docker.getManifestTemplateClass(), buildConfiguration.getTargetFormat());
+    Assert.assertSame(V22ManifestTemplate.class, buildConfiguration.getTargetFormat());
 
     Assert.assertEquals("jib-core", buildConfiguration.getToolName());
 
@@ -208,8 +208,7 @@ public class JibContainerBuilderTest {
                     .withAdditionalTag("tag2")
                     .setToolName("toolName"),
                 MoreExecutors.newDirectExecutorService());
-    Assert.assertSame(
-        ImageFormat.OCI.getManifestTemplateClass(), buildConfiguration.getTargetFormat());
+    Assert.assertSame(OCIManifestTemplate.class, buildConfiguration.getTargetFormat());
     Assert.assertEquals(
         ImmutableSet.of("latest", "tag1", "tag2"), buildConfiguration.getAllTargetImageTags());
     Assert.assertEquals("toolName", buildConfiguration.getToolName());
